Refactor proto module into focused submodules
