+++ AGENTS.md
@@ -143,6 +143,7 @@ metadata/update/DNS/TLS inspection modes, and executes requests via `src/http`.
 - Rust `--timing` enables DNS pre-resolution timing and transport connection timing so the waterfall includes DNS, TCP, TLS/QUIC, TTFB, and Body phases. The direct transport owns DNS/TCP/TLS/QUIC setup, so keep new timing instrumentation in `src/http/transport/` and `src/net.rs`.
 - `src/main.rs` runs the Tokio runtime on an explicitly larger stack thread, and the top-level app future in `src/app.rs` is heap-pinned before the shutdown-signal `tokio::select!`; do not move it back to `tokio::pin!` or the default process main stack because the combined async request/WebSocket/inspection state can overflow Windows' smaller main-thread stack even for metadata commands.
 - Rust response body paging is controlled by `--pager auto|on|off` or `pager = ...`; `auto` routes terminal stdout through `$PAGER` when set, otherwise falls back to `less -FIRX`, `on` forces the pager, and `off` disables it. `NO_PAGER` disables the default `auto` pager, and `$LESS` is honored by not adding fallback `less` flags when it is set. `$PAGER` is split with POSIX shell-style quoting but launched directly, so shell operators are not interpreted. Image responses and output-file writes bypass the pager.
+- Protocol buffer support is split under `src/proto/`: `schema.rs` owns descriptor loading, local schema lookup, and `protoc` invocation; `discovery.rs` owns local gRPC discovery rendering; `convert.rs` owns JSON/protobuf conversion and descriptor-backed gRPC frame formatting; `json_stream.rs` owns incremental JSON-to-gRPC request streaming and stdin/Windows pipe reading; `descriptor_wire.rs` owns small descriptor wire parsing helpers.
 - Timeout duration parsing, Go-style duration formatting, elapsed request budgets, connect/DNS timeout caps, and shared `request timed out after ...` errors live in `src/duration.rs`; HTTP, WebSocket, DNS inspection, and TLS inspection paths should reuse `TimeoutBudget` instead of recomputing remaining time locally, and response body deadlines should preserve the original request timeout diagnostic.
 - HTTP retries keep the original `--timeout` operation deadline across attempts, cap retry sleeps to the remaining budget, refresh per-request timeouts before each send, and use best-effort byte/time-bounded drains for retry cleanup responses.
 - Custom/direct DNS observes timeout budgets during transport setup: `--connect-timeout` bounds DNS and connect work when set, otherwise DNS uses the remaining `--timeout` budget, and DoH lookup clients receive the same budget.
