+++ src/proto/mod.rs
@@ -1,959 +1,23 @@
-use bytes::Bytes;
-use prost::Message;
-use prost_reflect::{
-    Cardinality, DescriptorPool, DynamicMessage, FieldDescriptor, Kind, MessageDescriptor,
-    MethodDescriptor, SerializeOptions, ServiceDescriptor,
+mod convert;
+mod descriptor_wire;
+mod discovery;
+mod json_stream;
+mod schema;
+
+pub(crate) use convert::grpc_request_body;
+pub use convert::{
+    format_grpc_frame_with_descriptor, format_grpc_stream_with_descriptor, json_to_protobuf,
+    protobuf_to_json, protobuf_to_json_compact,
 };
-use std::io::Read;
-use std::path::{Path, PathBuf};
-use std::process::Command;
-use std::time::{SystemTime, UNIX_EPOCH};
-use tokio::io::{AsyncRead, AsyncReadExt};
-
-use crate::error::FetchError;
-use crate::grpc::encoding::{self, MessageEncoding};
-use crate::grpc::framing;
-
-const MAX_GRPC_JSON_PENDING_BYTES: usize = framing::MAX_MESSAGE_SIZE;
-
-#[derive(Debug, Clone)]
-pub struct Schema {
-    pool: DescriptorPool,
-}
-
-impl Schema {
-    pub fn from_descriptor_set(bytes: &[u8]) -> Result<Self, ProtoError> {
-        let pool = DescriptorPool::decode(bytes).map_err(|err| {
-            ProtoError::Message(format!("failed to create file descriptors: {err}"))
-        })?;
-        Ok(Self { pool })
-    }
-
-    pub fn from_file_descriptor_protos(files: &[Vec<u8>]) -> Result<Self, ProtoError> {
-        let mut pool = DescriptorPool::new();
-        let mut seen = std::collections::BTreeSet::new();
-        for file in files {
-            let name = reflected_file_name(file)?;
-            if !seen.insert(name) {
-                continue;
-            }
-            pool.decode_file_descriptor_proto(file.as_slice())
-                .map_err(|err| {
-                    ProtoError::Message(format!("failed to decode reflected descriptor: {err}"))
-                })?;
-        }
-        Ok(Self { pool })
-    }
-
-    pub fn load_descriptor_set_file(path: &str) -> Result<Self, FetchError> {
-        let bytes = std::fs::read(path).map_err(|err| {
-            FetchError::Message(format!("failed to read descriptor set file: {err}"))
-        })?;
-        Self::from_descriptor_set(&bytes).map_err(|err| FetchError::Message(err.to_string()))
-    }
-
-    pub fn find_method(&self, full_name: &str) -> Result<MethodDescriptor, ProtoError> {
-        let (service_name, method_name) = split_method_name(full_name)?;
-        let service = self
-            .pool
-            .get_service_by_name(service_name)
-            .ok_or_else(|| ProtoError::Message(format!("service not found: {service_name}")))?;
-        let method = service
-            .methods()
-            .find(|method| method.name() == method_name);
-        method.ok_or_else(|| {
-            ProtoError::Message(format!(
-                "method {method_name} not found in service {service_name}"
-            ))
-        })
-    }
-
-    pub fn find_service(&self, name: &str) -> Option<ServiceDescriptor> {
-        self.pool.get_service_by_name(normalize_symbol_name(name))
-    }
-
-    pub fn find_message(&self, name: &str) -> Option<MessageDescriptor> {
-        self.pool.get_message_by_name(normalize_symbol_name(name))
-    }
-
-    pub fn messages(&self) -> Vec<String> {
-        let mut messages: Vec<_> = self
-            .pool
-            .all_messages()
-            .map(|message| message.full_name().to_string())
-            .collect();
-        messages.sort();
-        messages
-    }
-
-    pub fn services(&self) -> Vec<String> {
-        let mut services: Vec<_> = self
-            .pool
-            .services()
-            .map(|service| service.full_name().to_string())
-            .collect();
-        services.sort();
-        services
-    }
-}
-
-pub fn execute_local_discovery(cli: &crate::cli::Cli) -> Result<i32, FetchError> {
-    let Some(schema) = load_local_schema(cli)? else {
-        return Err("gRPC discovery requires --proto-file or --proto-desc".into());
-    };
-
-    if cli.grpc_list {
-        crate::core::write_stdout(service_list_output(schema.services()))?;
-        return Ok(0);
-    }
-
-    if let Some(symbol) = cli.grpc_describe.as_deref() {
-        let symbol = normalize_symbol_name(symbol);
-        crate::core::write_stdout(describe_symbol(&schema, symbol)?)?;
-        return Ok(0);
-    }
-
-    Err("gRPC discovery requires --grpc-list or --grpc-describe".into())
-}
-
-fn service_list_output(services: Vec<String>) -> String {
-    let mut output = String::new();
-    for service in services {
-        output.push_str(&service);
-        output.push('\n');
-    }
-    output
-}
-
-pub fn load_local_schema(cli: &crate::cli::Cli) -> Result<Option<Schema>, FetchError> {
-    let proto_files = proto_file_paths(&cli.proto_files);
-    if !proto_files.is_empty() {
-        return compile_protos(&proto_files, &cli.proto_imports)
-            .map(Some)
-            .map_err(|err| FetchError::Message(err.to_string()));
-    }
-    cli.proto_desc
-        .as_deref()
-        .map(Schema::load_descriptor_set_file)
-        .transpose()
-}
-
-pub fn proto_file_paths(values: &[String]) -> Vec<String> {
-    values
-        .iter()
-        .flat_map(|value| value.split(','))
-        .map(str::trim)
-        .filter(|path| !path.is_empty())
-        .map(ToOwned::to_owned)
-        .collect()
-}
-
-pub fn compile_protos(
-    proto_files: &[String],
-    import_paths: &[String],
-) -> Result<Schema, ProtoError> {
-    let descriptor_path = TempDescriptorSet::create()?;
-    let mut command = Command::new("protoc");
-    command
-        .arg(format!(
-            "--descriptor_set_out={}",
-            descriptor_path.path().display()
-        ))
-        .arg("--include_imports");
-
-    if import_paths.is_empty() {
-        for dir in default_proto_import_paths(proto_files)? {
-            command.arg(format!("-I={}", dir.display()));
-        }
-    } else {
-        for import in import_paths {
-            command.arg(format!("-I={import}"));
-        }
-    }
-    for file in proto_files {
-        command.arg(file);
-    }
-
-    let output = command.output().map_err(|err| {
-        if err.kind() == std::io::ErrorKind::NotFound {
-            ProtoError::ProtocNotFound
-        } else {
-            ProtoError::Message(format!("failed to run protoc: {err}"))
-        }
-    })?;
-    if !output.status.success() {
-        let message = String::from_utf8_lossy(&output.stderr).trim().to_string();
-        let message = if message.is_empty() {
-            exit_status_message(output.status)
-        } else {
-            message
-        };
-        return Err(ProtoError::Protoc(message));
-    }
-
-    let bytes = std::fs::read(descriptor_path.path())
-        .map_err(|err| ProtoError::Message(format!("failed to read descriptor set file: {err}")))?;
-    Schema::from_descriptor_set(&bytes)
-}
-
-pub fn method_for_url(schema: &Schema, url: &url::Url) -> Result<MethodDescriptor, FetchError> {
-    schema
-        .find_method(url.path().trim_start_matches('/'))
-        .map_err(|err| FetchError::Message(err.to_string()))
-}
-
-pub(crate) fn grpc_request_body(
-    body: crate::http::RequestBody,
-    method: Option<&MethodDescriptor>,
-) -> Result<crate::http::RequestBody, FetchError> {
-    let Some(method) = method else {
-        return frame_raw_body(body);
-    };
-
-    if method.is_client_streaming() {
-        let Some(body) = body else {
-            return Ok(None);
-        };
-        return Ok(Some(
-            crate::http::RequestBodyPayload::from_grpc_json_stream(
-                body,
-                method.input(),
-                Some(grpc_content_type()),
-            ),
-        ));
-    }
-
-    let limit_error = grpc_request_body_limit_error();
-    let raw = if let Some((bytes, _)) =
-        crate::http::request_body_into_bytes_limited(body, framing::MAX_MESSAGE_SIZE, &limit_error)?
-    {
-        json_to_protobuf(&bytes, &method.input())
-            .map_err(|err| FetchError::Message(err.to_string()))?
-    } else {
-        Vec::new()
-    };
-    Ok(Some(crate::http::RequestBodyPayload::from_bytes(
-        framing::frame(&raw, false).map_err(|err| FetchError::Message(err.to_string()))?,
-        Some(grpc_content_type()),
-    )))
-}
-
-pub fn format_grpc_stream_with_descriptor(
-    bytes: &[u8],
-    desc: &MessageDescriptor,
-    message_encoding: &MessageEncoding,
-) -> Result<String, ProtoError> {
-    let frames = framing::read_frames(bytes)
-        .map_err(|err| ProtoError::Message(format!("failed to read gRPC stream: {err}")))?;
-    let mut out = String::new();
-    let mut wrote_any = false;
-    for frame in &frames {
-        let formatted = format_grpc_frame_with_descriptor(frame, desc, message_encoding)?;
-        if wrote_any && !out.ends_with('\n') {
-            out.push('\n');
-        }
-        out.push_str(&formatted);
-        wrote_any = true;
-    }
-    Ok(out)
-}
-
-pub fn format_grpc_frame_with_descriptor(
-    frame: &framing::Frame,
-    desc: &MessageDescriptor,
-    message_encoding: &MessageEncoding,
-) -> Result<String, ProtoError> {
-    let data = encoding::decompress_frame(frame, message_encoding)
-        .map_err(|err| ProtoError::Message(err.to_string()))?;
-    let msg = decode_dynamic_message(data.as_slice(), desc)?;
-    let mut formatted = String::from_utf8(serialize_dynamic_message_json(&msg, true)?)
-        .map_err(|err| ProtoError::Message(format!("failed to format protobuf JSON: {err}")))?;
-    formatted.push('\n');
-    Ok(formatted)
-}
-
-pub fn describe_symbol(schema: &Schema, symbol: &str) -> Result<String, FetchError> {
-    let symbol = normalize_symbol_name(symbol);
-    if symbol.contains('/') {
-        let method = schema
-            .find_method(symbol)
-            .map_err(|_| FetchError::Message(format!("symbol not found: {symbol}")))?;
-        return Ok(render_method_description(&method));
-    }
-
-    if let Some(service) = schema.find_service(symbol) {
-        return Ok(render_service_description(&service));
-    }
-    if let Ok(method) = schema.find_method(symbol) {
-        return Ok(render_method_description(&method));
-    }
-    if let Some(message) = schema.find_message(symbol) {
-        return Ok(render_message_description(&message));
-    }
-    Err(format!("symbol not found: {symbol}").into())
-}
-
-pub fn normalize_symbol_name(symbol: &str) -> &str {
-    symbol.trim_start_matches(['/', '.'])
-}
-
-fn render_service_description(service: &ServiceDescriptor) -> String {
-    let mut out = format!("service {}\n", service.full_name());
-    for method in service.methods() {
-        out.push('\n');
-        out.push_str(method.name());
-        out.push('\n');
-        out.push_str("  rpc: ");
-        out.push_str(rpc_type(&method));
-        out.push('\n');
-        out.push_str("  request: ");
-        out.push_str(method.input().full_name());
-        out.push('\n');
-        out.push_str("  response: ");
-        out.push_str(method.output().full_name());
-        out.push('\n');
-    }
-    out
-}
-
-fn render_method_description(method: &MethodDescriptor) -> String {
-    let mut out = format!(
-        "method {}/{}\n",
-        method.parent_service().full_name(),
-        method.name()
-    );
-    out.push_str("rpc: ");
-    out.push_str(rpc_type(method));
-    out.push('\n');
-    out.push_str("request: ");
-    out.push_str(method.input().full_name());
-    out.push('\n');
-    out.push_str("response: ");
-    out.push_str(method.output().full_name());
-    out.push('\n');
-    out
-}
-
-fn render_message_description(message: &MessageDescriptor) -> String {
-    let mut out = format!("message {}\n", message.full_name());
-    for field in message.fields() {
-        out.push('\n');
-        out.push_str(&format!(
-            "{}  {}  {}  {}\n",
-            field.number(),
-            field.name(),
-            field_label(&field),
-            field_type(&field)
-        ));
-    }
-    out
-}
-
-fn rpc_type(method: &MethodDescriptor) -> &'static str {
-    match (method.is_client_streaming(), method.is_server_streaming()) {
-        (true, true) => "bidi-stream",
-        (true, false) => "client-stream",
-        (false, true) => "server-stream",
-        (false, false) => "unary",
-    }
-}
-
-fn field_label(field: &FieldDescriptor) -> &'static str {
-    if field.is_list() {
-        return "repeated";
-    }
-    match field.cardinality() {
-        Cardinality::Required => "required",
-        Cardinality::Optional => "optional",
-        Cardinality::Repeated => "repeated",
-    }
-}
-
-fn field_type(field: &FieldDescriptor) -> String {
-    match field.kind() {
-        Kind::Message(message) => message.full_name().to_string(),
-        Kind::Enum(en) => en.full_name().to_string(),
-        Kind::Double => "double".to_string(),
-        Kind::Float => "float".to_string(),
-        Kind::Int32 => "int32".to_string(),
-        Kind::Int64 => "int64".to_string(),
-        Kind::Uint32 => "uint32".to_string(),
-        Kind::Uint64 => "uint64".to_string(),
-        Kind::Sint32 => "sint32".to_string(),
-        Kind::Sint64 => "sint64".to_string(),
-        Kind::Fixed32 => "fixed32".to_string(),
-        Kind::Fixed64 => "fixed64".to_string(),
-        Kind::Sfixed32 => "sfixed32".to_string(),
-        Kind::Sfixed64 => "sfixed64".to_string(),
-        Kind::Bool => "bool".to_string(),
-        Kind::String => "string".to_string(),
-        Kind::Bytes => "bytes".to_string(),
-    }
-}
-
-fn frame_raw_body(body: crate::http::RequestBody) -> Result<crate::http::RequestBody, FetchError> {
-    let limit_error = grpc_request_body_limit_error();
-    let raw = crate::http::request_body_into_bytes_limited(
-        body,
-        framing::MAX_MESSAGE_SIZE,
-        &limit_error,
-    )?
-    .map(|(bytes, _)| bytes)
-    .unwrap_or_default();
-    Ok(Some(crate::http::RequestBodyPayload::from_bytes(
-        framing::frame(&raw, false).map_err(|err| FetchError::Message(err.to_string()))?,
-        Some(grpc_content_type()),
-    )))
-}
-
-fn grpc_request_body_limit_error() -> String {
-    format!(
-        "gRPC request body exceeds maximum of {} bytes",
-        framing::MAX_MESSAGE_SIZE
-    )
-}
-
-pub fn json_to_protobuf(json: &[u8], desc: &MessageDescriptor) -> Result<Vec<u8>, ProtoError> {
-    let mut deserializer = serde_json::Deserializer::from_slice(json);
-    let options = prost_reflect::DeserializeOptions::new().deny_unknown_fields(false);
-    let msg = DynamicMessage::deserialize_with_options(desc.clone(), &mut deserializer, &options)
-        .map_err(|err| ProtoError::Message(err.to_string()))?;
-    deserializer
-        .end()
-        .map_err(|err| ProtoError::Message(err.to_string()))?;
-    Ok(msg.encode_to_vec())
-}
-
-pub fn protobuf_to_json(bytes: &[u8], desc: &MessageDescriptor) -> Result<Vec<u8>, ProtoError> {
-    let msg = decode_dynamic_message(bytes, desc)?;
-    serialize_dynamic_message_json(&msg, true)
-}
-
-pub fn protobuf_to_json_compact(
-    bytes: &[u8],
-    desc: &MessageDescriptor,
-) -> Result<Vec<u8>, ProtoError> {
-    let msg = decode_dynamic_message(bytes, desc)?;
-    serialize_dynamic_message_json(&msg, false)
-}
-
-fn decode_dynamic_message(
-    bytes: &[u8],
-    desc: &MessageDescriptor,
-) -> Result<DynamicMessage, ProtoError> {
-    DynamicMessage::decode(desc.clone(), bytes)
-        .map_err(|err| ProtoError::Message(format!("failed to decode protobuf message: {err}")))
-}
-
-fn serialize_dynamic_message_json(
-    msg: &DynamicMessage,
-    pretty: bool,
-) -> Result<Vec<u8>, ProtoError> {
-    let options = SerializeOptions::new().use_proto_field_name(true);
-    if pretty {
-        let mut serializer = serde_json::Serializer::pretty(Vec::new());
-        msg.serialize_with_options(&mut serializer, &options)
-            .map_err(|err| ProtoError::Message(format!("failed to format protobuf JSON: {err}")))?;
-        return Ok(serializer.into_inner());
-    }
-
-    let mut serializer = serde_json::Serializer::new(Vec::new());
-    msg.serialize_with_options(&mut serializer, &options)
-        .map_err(|err| ProtoError::Message(format!("failed to format protobuf JSON: {err}")))?;
-    Ok(serializer.into_inner())
-}
-
-pub fn stream_json_to_grpc_frames(
-    json_stream: &[u8],
-    desc: &MessageDescriptor,
-) -> Result<Vec<u8>, ProtoError> {
-    let mut out = Vec::new();
-    let stream = serde_json::Deserializer::from_slice(json_stream).into_iter::<serde_json::Value>();
-    for value in stream {
-        let value = value
-            .map_err(|err| ProtoError::Message(format!("failed to decode JSON message: {err}")))?;
-        out.extend_from_slice(&json_value_to_grpc_frame(value, desc)?);
-    }
-    Ok(out)
-}
-
-pub(crate) fn json_reader_to_grpc_frame_stream<R>(
-    reader: R,
-    desc: MessageDescriptor,
-) -> impl futures_util::Stream<Item = Result<Bytes, std::io::Error>>
-where
-    R: AsyncRead + Unpin + Send + 'static,
-{
-    json_reader_to_grpc_frame_stream_with_limit(reader, desc, MAX_GRPC_JSON_PENDING_BYTES)
-}
-
-fn json_reader_to_grpc_frame_stream_with_limit<R>(
-    reader: R,
-    desc: MessageDescriptor,
-    max_pending_json_bytes: usize,
-) -> impl futures_util::Stream<Item = Result<Bytes, std::io::Error>>
-where
-    R: AsyncRead + Unpin + Send + 'static,
-{
-    struct State<R> {
-        reader: R,
-        desc: MessageDescriptor,
-        pending: Vec<u8>,
-        eof: bool,
-        max_pending_json_bytes: usize,
-    }
-
-    futures_util::stream::try_unfold(
-        State {
-            reader,
-            desc,
-            pending: Vec::new(),
-            eof: false,
-            max_pending_json_bytes,
-        },
-        |mut state| async move {
-            loop {
-                discard_leading_json_whitespace(&mut state.pending);
-                validate_grpc_json_pending_len(state.pending.len(), state.max_pending_json_bytes)
-                    .map_err(std::io::Error::other)?;
-                match parse_next_json_value(&state.pending, state.eof) {
-                    Ok(JsonParse::Message { value, consumed }) => {
-                        state.pending.drain(..consumed);
-                        let frame = json_value_to_grpc_frame(value, &state.desc)
-                            .map_err(std::io::Error::other)?;
-                        return Ok(Some((Bytes::from(frame), state)));
-                    }
-                    Ok(JsonParse::NeedMore) => {}
-                    Ok(JsonParse::Done) => return Ok(None),
-                    Err(err) => return Err(std::io::Error::other(err)),
-                }
-
-                if state.eof {
-                    return Ok(None);
-                }
-
-                let mut buf = [0_u8; 16 * 1024];
-                let n = state.reader.read(&mut buf).await?;
-                if n == 0 {
-                    state.eof = true;
-                } else {
-                    state.pending.extend_from_slice(&buf[..n]);
-                }
-            }
-        },
-    )
-}
-
-pub(crate) fn stdin_json_to_grpc_frame_stream(
-    desc: MessageDescriptor,
-) -> impl futures_util::Stream<Item = Result<Bytes, std::io::Error>> {
-    const READ_BUF_LEN: usize = 16 * 1024;
-
-    let (tx, rx) = tokio::sync::mpsc::channel(1);
-    std::thread::spawn(move || {
-        let stdin = std::io::stdin();
-        let mut stdin_read = StdinReadState::new(&stdin);
-        let mut reader = stdin.lock();
-        let mut pending = Vec::new();
-        let mut eof = false;
-        let mut buf = [0_u8; READ_BUF_LEN];
-
-        loop {
-            discard_leading_json_whitespace(&mut pending);
-            if let Err(err) =
-                validate_grpc_json_pending_len(pending.len(), MAX_GRPC_JSON_PENDING_BYTES)
-            {
-                let _ = tx.blocking_send(Err(std::io::Error::other(err)));
-                break;
-            }
-            match parse_next_json_value(&pending, eof) {
-                Ok(JsonParse::Message { value, consumed }) => {
-                    pending.drain(..consumed);
-                    let item = json_value_to_grpc_frame(value, &desc)
-                        .map(Bytes::from)
-                        .map_err(std::io::Error::other);
-                    if tx.blocking_send(item).is_err() {
-                        break;
-                    }
-                }
-                Ok(JsonParse::NeedMore) => {}
-                Ok(JsonParse::Done) => break,
-                Err(err) => {
-                    let _ = tx.blocking_send(Err(std::io::Error::other(err)));
-                    break;
-                }
-            }
-
-            if eof {
-                break;
-            }
-
-            match stdin_read.read_chunk(&mut reader, &mut buf) {
-                Ok(0) => eof = true,
-                Ok(n) => pending.extend_from_slice(&buf[..n]),
-                Err(err) => {
-                    let _ = tx.blocking_send(Err(err));
-                    break;
-                }
-            }
-        }
-    });
-
-    futures_util::stream::unfold(rx, |mut rx| async {
-        rx.recv().await.map(|item| (item, rx))
-    })
-}
-
-fn discard_leading_json_whitespace(pending: &mut Vec<u8>) {
-    match pending.iter().position(|byte| !byte.is_ascii_whitespace()) {
-        Some(0) => {}
-        Some(first_non_whitespace) => {
-            pending.drain(..first_non_whitespace);
-        }
-        None => pending.clear(),
-    }
-}
-
-fn validate_grpc_json_pending_len(len: usize, max: usize) -> Result<(), ProtoError> {
-    if len > max {
-        return Err(ProtoError::Message(format!(
-            "gRPC JSON message exceeds {max} bytes before a complete JSON value"
-        )));
-    }
-    Ok(())
-}
-
-#[cfg(not(windows))]
-struct StdinReadState;
-
-#[cfg(not(windows))]
-impl StdinReadState {
-    fn new(_stdin: &std::io::Stdin) -> Self {
-        Self
-    }
-
-    fn read_chunk<R: Read>(&mut self, reader: &mut R, buf: &mut [u8]) -> std::io::Result<usize> {
-        reader.read(buf)
-    }
-}
-
-#[cfg(windows)]
-struct StdinReadState {
-    handle: windows_sys::Win32::Foundation::HANDLE,
-    is_pipe: bool,
-}
-
-#[cfg(windows)]
-impl StdinReadState {
-    fn new(stdin: &std::io::Stdin) -> Self {
-        use std::os::windows::io::AsRawHandle;
-        use windows_sys::Win32::Storage::FileSystem::{FILE_TYPE_PIPE, GetFileType};
-
-        let handle = stdin.as_raw_handle();
-        // SAFETY: the handle is borrowed from stdio and remains valid while
-        // stdin is alive; GetFileType does not retain it.
-        let is_pipe = unsafe { GetFileType(handle) } == FILE_TYPE_PIPE;
-        Self { handle, is_pipe }
-    }
-
-    fn read_chunk<R: Read>(&mut self, reader: &mut R, buf: &mut [u8]) -> std::io::Result<usize> {
-        if !self.is_pipe {
-            return reader.read(buf);
-        }
-
-        loop {
-            match self.pipe_bytes_available()? {
-                PipeReadiness::Available(0) => {
-                    std::thread::sleep(std::time::Duration::from_millis(1));
-                }
-                PipeReadiness::Available(n) => {
-                    let len = n.min(buf.len());
-                    return reader.read(&mut buf[..len]);
-                }
-                PipeReadiness::Eof => return Ok(0),
-            }
-        }
-    }
-
-    fn pipe_bytes_available(&self) -> std::io::Result<PipeReadiness> {
-        use windows_sys::Win32::Foundation::{ERROR_BROKEN_PIPE, ERROR_HANDLE_EOF, GetLastError};
-        use windows_sys::Win32::System::Pipes::PeekNamedPipe;
-
-        let mut available = 0_u32;
-        // SAFETY: the handle is a stdin pipe handle. All optional output
-        // pointers are either null or valid for the duration of the call.
-        let ok = unsafe {
-            PeekNamedPipe(
-                self.handle,
-                std::ptr::null_mut(),
-                0,
-                std::ptr::null_mut(),
-                &mut available,
-                std::ptr::null_mut(),
-            )
-        };
-        if ok != 0 {
-            return Ok(PipeReadiness::Available(available as usize));
-        }
-
-        // SAFETY: GetLastError reads the thread-local error from the failed
-        // PeekNamedPipe call above.
-        match unsafe { GetLastError() } {
-            ERROR_BROKEN_PIPE | ERROR_HANDLE_EOF => Ok(PipeReadiness::Eof),
-            _ => Err(std::io::Error::last_os_error()),
-        }
-    }
-}
-
-#[cfg(windows)]
-enum PipeReadiness {
-    Available(usize),
-    Eof,
-}
-
-enum JsonParse {
-    Message {
-        value: serde_json::Value,
-        consumed: usize,
-    },
-    NeedMore,
-    Done,
-}
-
-fn parse_next_json_value(buf: &[u8], eof: bool) -> Result<JsonParse, ProtoError> {
-    if buf.is_empty() {
-        return Ok(if eof {
-            JsonParse::Done
-        } else {
-            JsonParse::NeedMore
-        });
-    }
-    if buf.iter().all(u8::is_ascii_whitespace) {
-        return Ok(if eof {
-            JsonParse::Done
-        } else {
-            JsonParse::NeedMore
-        });
-    }
-
-    let mut stream = serde_json::Deserializer::from_slice(buf).into_iter::<serde_json::Value>();
-    match stream.next() {
-        Some(Ok(value)) => Ok(JsonParse::Message {
-            value,
-            consumed: stream.byte_offset(),
-        }),
-        Some(Err(err)) if err.is_eof() && !eof => Ok(JsonParse::NeedMore),
-        Some(Err(err)) => Err(ProtoError::Message(format!(
-            "failed to decode JSON message: {err}"
-        ))),
-        None => Ok(if eof {
-            JsonParse::Done
-        } else {
-            JsonParse::NeedMore
-        }),
-    }
-}
-
-fn json_value_to_grpc_frame(
-    value: serde_json::Value,
-    desc: &MessageDescriptor,
-) -> Result<Vec<u8>, ProtoError> {
-    let value_text = value.to_string();
-    let mut deserializer = serde_json::Deserializer::from_str(&value_text);
-    let options = prost_reflect::DeserializeOptions::new().deny_unknown_fields(false);
-    let msg =
-        match DynamicMessage::deserialize_with_options(desc.clone(), &mut deserializer, &options) {
-            Ok(msg) => msg,
-            Err(err) => {
-                return Err(ProtoError::Message(format!(
-                    "failed to convert JSON to protobuf: {err}"
-                )));
-            }
-        };
-    deserializer
-        .end()
-        .map_err(|err| ProtoError::Message(format!("failed to decode JSON message: {err}")))?;
-    framing::frame(&msg.encode_to_vec(), false).map_err(|err| ProtoError::Message(err.to_string()))
-}
-
-fn split_method_name(full_name: &str) -> Result<(&str, &str), ProtoError> {
-    let full_name = normalize_symbol_name(full_name);
-    if let Some((service, method)) = full_name.rsplit_once('/') {
-        if !service.is_empty() && !method.is_empty() {
-            return Ok((service, method));
-        }
-    } else if let Some((service, method)) = full_name.rsplit_once('.')
-        && !service.is_empty()
-        && !method.is_empty()
-    {
-        return Ok((service, method));
-    }
-    Err(ProtoError::Message(format!(
-        "invalid method name format: {full_name} (expected 'Service/Method' or 'Service.Method')"
-    )))
-}
-
-fn grpc_content_type() -> String {
-    crate::grpc::headers::PROTO_CONTENT_TYPE.to_string()
-}
-
-fn reflected_file_name(file: &[u8]) -> Result<String, ProtoError> {
-    let mut raw = file;
-    while !raw.is_empty() {
-        let (field, wire) = read_key(&mut raw)?;
-        if field == 1 && wire == 2 {
-            return read_len_string(&mut raw);
-        }
-        skip_wire_value(wire, &mut raw)?;
-    }
-    Err(ProtoError::Message(
-        "reflected descriptor is missing a file name".to_string(),
-    ))
-}
-
-fn default_proto_import_paths(proto_files: &[String]) -> Result<Vec<PathBuf>, ProtoError> {
-    let cwd = std::env::current_dir()
-        .map_err(|err| ProtoError::Message(format!("failed to get current directory: {err}")))?;
-    let mut seen = std::collections::HashSet::new();
-    let mut dirs = Vec::new();
-    for file in proto_files {
-        let path = Path::new(file);
-        let dir = path
-            .parent()
-            .filter(|dir| !dir.as_os_str().is_empty())
-            .unwrap_or_else(|| Path::new("."));
-        let abs = if dir.is_absolute() {
-            dir.to_path_buf()
-        } else {
-            cwd.join(dir)
-        };
-        if seen.insert(abs.clone()) {
-            dirs.push(abs);
-        }
-    }
-    Ok(dirs)
-}
-
-fn exit_status_message(status: std::process::ExitStatus) -> String {
-    match status.code() {
-        Some(code) => format!("exit status {code}"),
-        None => "process terminated by signal".to_string(),
-    }
-}
-
-struct TempDescriptorSet {
-    path: PathBuf,
-}
-
-impl TempDescriptorSet {
-    fn create() -> Result<Self, ProtoError> {
-        let temp_dir = std::env::temp_dir();
-        let pid = std::process::id();
-        let nanos = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .map(|duration| duration.as_nanos())
-            .unwrap_or_default();
-        for attempt in 0..100 {
-            let path = temp_dir.join(format!("fetch-proto-{pid}-{nanos}-{attempt}.pb"));
-            match std::fs::OpenOptions::new()
-                .write(true)
-                .create_new(true)
-                .open(&path)
-            {
-                Ok(file) => {
-                    drop(file);
-                    return Ok(Self { path });
-                }
-                Err(err) if err.kind() == std::io::ErrorKind::AlreadyExists => continue,
-                Err(err) => {
-                    return Err(ProtoError::Message(format!(
-                        "failed to create temp file: {err}"
-                    )));
-                }
-            }
-        }
-        Err(ProtoError::Message(
-            "failed to create temp file: too many collisions".to_string(),
-        ))
-    }
-
-    fn path(&self) -> &Path {
-        &self.path
-    }
-}
-
-impl Drop for TempDescriptorSet {
-    fn drop(&mut self) {
-        let _ = std::fs::remove_file(&self.path);
-    }
-}
-
-fn read_key(raw: &mut &[u8]) -> Result<(u64, u8), ProtoError> {
-    let key = read_varint(raw)?;
-    Ok((key >> 3, (key & 0x07) as u8))
-}
-
-fn read_varint(raw: &mut &[u8]) -> Result<u64, ProtoError> {
-    let mut value = 0_u64;
-    for shift in (0..64).step_by(7) {
-        let Some((&byte, rest)) = raw.split_first() else {
-            return Err(ProtoError::Message(
-                "unexpected EOF while reading varint".to_string(),
-            ));
-        };
-        *raw = rest;
-        value |= u64::from(byte & 0x7f) << shift;
-        if byte & 0x80 == 0 {
-            return Ok(value);
-        }
-    }
-    Err(ProtoError::Message("varint overflows uint64".to_string()))
-}
-
-fn read_len_bytes(raw: &mut &[u8]) -> Result<Vec<u8>, ProtoError> {
-    let len = usize::try_from(read_varint(raw)?)
-        .map_err(|_| ProtoError::Message("length overflows usize".to_string()))?;
-    if raw.len() < len {
-        return Err(ProtoError::Message(
-            "unexpected EOF while reading bytes".to_string(),
-        ));
-    }
-    let out = raw[..len].to_vec();
-    *raw = &raw[len..];
-    Ok(out)
-}
-
-fn read_len_string(raw: &mut &[u8]) -> Result<String, ProtoError> {
-    String::from_utf8(read_len_bytes(raw)?)
-        .map_err(|err| ProtoError::Message(format!("invalid UTF-8 string: {err}")))
-}
-
-fn skip_wire_value(wire: u8, raw: &mut &[u8]) -> Result<(), ProtoError> {
-    match wire {
-        0 => {
-            read_varint(raw)?;
-        }
-        1 => skip_fixed(raw, 8)?,
-        2 => {
-            let len = usize::try_from(read_varint(raw)?)
-                .map_err(|_| ProtoError::Message("length overflows usize".to_string()))?;
-            skip_fixed(raw, len)?;
-        }
-        5 => skip_fixed(raw, 4)?,
-        _ => return Err(ProtoError::Message(format!("unsupported wire type {wire}"))),
-    }
-    Ok(())
-}
-
-fn skip_fixed(raw: &mut &[u8], len: usize) -> Result<(), ProtoError> {
-    if raw.len() < len {
-        return Err(ProtoError::Message(
-            "unexpected EOF while skipping field".to_string(),
-        ));
-    }
-    *raw = &raw[len..];
-    Ok(())
-}
+pub use discovery::{describe_symbol, execute_local_discovery};
+#[cfg(test)]
+pub(crate) use json_stream::json_reader_to_grpc_frame_stream_with_limit;
+pub use json_stream::stream_json_to_grpc_frames;
+pub(crate) use json_stream::{json_reader_to_grpc_frame_stream, stdin_json_to_grpc_frame_stream};
+pub use schema::{
+    Schema, compile_protos, load_local_schema, method_for_url, normalize_symbol_name,
+    proto_file_paths,
+};
 
 #[derive(Debug, thiserror::Error)]
 pub enum ProtoError {
@@ -970,9 +34,13 @@ pub enum ProtoError {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::grpc::encoding::MessageEncoding;
+    use crate::grpc::framing;
     use flate2::Compression;
     use flate2::write::GzEncoder;
     use futures_util::TryStreamExt;
+    use prost::Message;
+    use prost_reflect::MessageDescriptor;
     use prost_types::{
         DescriptorProto, EnumDescriptorProto, EnumValueDescriptorProto, FieldDescriptorProto,
         FileDescriptorProto, FileDescriptorSet, MessageOptions, MethodDescriptorProto,
